refactor(cleanup): adhere to naming conventions
