+++ src/containers/home/home-greeting.js
@@ -7,17 +7,13 @@ export const HomeGreeting = () => {
   const pinnedTopics = useSelector((state) => state.settings.pinnedTopics)
   const timeOfDay = getTimeOfDay()
 
-  const greeting = firstName
-    ? `Good ${timeOfDay}, ${firstName}!`
-    : `Good ${timeOfDay}!`
-  const description = (pinnedTopics?.length === 0)
-    ? `Select topics to see popular articles and our editors’ top picks.`
-    : null
+  const showName = firstName && firstName.length < 15
+  const greeting = showName ? `Good ${timeOfDay}, ${firstName}!` : `Good ${timeOfDay}!`
 
-  return (
-    <CardPageHeader
-      title={greeting}
-      subHeading={description}
-    />
-  )
+  const description =
+    pinnedTopics?.length === 0
+      ? `Select topics to see popular articles and our editors’ top picks.`
+      : null
+
+  return <CardPageHeader title={greeting} subHeading={description} />
 }
